% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/dangas_a.r
\name{extract_a}
\alias{extract_a}
\title{Helps to extract a substring present within two delimiters.}
\usage{
extract_a(start, end, string, delim)
}
\arguments{
\item{start}{point from which the matching starts}

\item{end}{point where matching ends}

\item{string}{input string}

\item{delim}{Turn \code{delim = TRUE} to include also the start and end delimiters in the final output.}
}
\value{
substring present in-between the start and end delimiters.
}
\description{
\code{extract_a} returns the substring present in-between the start and end delimiters.
Note that it would return only the first occurance.
}
\examples{
extract_a("<?", "?>", "foo <?php info?> bar")
extract_a("<?", "?>", "foo <?php info?> bar", delim = TRUE)
}

