% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/dangas_a.r
\name{extract_all_a}
\alias{extract_all_a}
\title{Helps to extract all the substrings present within two delimiters.}
\usage{
extract_all_a(start, end, string, delim)
}
\arguments{
\item{start}{point from which the matching starts.}

\item{end}{point where matching ends.}

\item{string}{input string.}

\item{delim}{Turn \code{delim = TRUE} to include also the start and end delimiters in the final output.}
}
\value{
substring present in-between the from and to delimiters. Note that it would return all the occurances. Final output would be in the form of list.
}
\description{
\code{extract_all_a} returns all the substrings present in-between the start and end delimiters.
Note that it would return all the occurances.
}
\examples{
extract_all_a("<?", "?>", "foo <?php info?> bar <?r info?>")
extract_all_a("<?", "?>", "foo <?php info?> bar <?r info?>", delim = TRUE)
}

